move TracingListenerFactory
