+++ elastic-job-lite-tracing/elastic-job-lite-tracing-rdb/src/main/java/org/apache/shardingsphere/elasticjob/lite/tracing/rdb/config/RDBTracingListenerConfiguration.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.elasticjob.lite.tracing.rdb.config;
 
-import org.apache.shardingsphere.elasticjob.lite.tracing.config.TracingListenerConfiguration;
+import org.apache.shardingsphere.elasticjob.lite.tracing.listener.TracingListenerConfiguration;
 import org.apache.shardingsphere.elasticjob.lite.tracing.exception.TracingConfigurationException;
 import org.apache.shardingsphere.elasticjob.lite.tracing.listener.TracingListener;
 import org.apache.shardingsphere.elasticjob.lite.tracing.rdb.listener.RDBTracingListener;
